# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"GreenCredits API (MVP)","version":"0.1.0"},"paths":{"/auth/register":{"post":{"tags":["Auth"],"summary":"Register","operationId":"register_auth_register_post","responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthTokenResponse"}}}}}}},"/auth/login":{"post":{"tags":["Auth"],"summary":"Login","operationId":"login_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthLoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthTokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/wallet/balance":{"get":{"tags":["Wallet"],"summary":"Balance","operationId":"balance_wallet_balance_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WalletBalanceResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/wallet/history":{"get":{"tags":["Wallet"],"summary":"History","operationId":"history_wallet_history_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Pagesize"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/WalletHistoryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/subscriptions/me":{"get":{"tags":["Subscriptions"],"summary":"Get Me","operationId":"get_me_subscriptions_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subscription"}}}}},"security":[{"HTTPBearer":[]}]}},"/subscriptions/activate":{"post":{"tags":["Subscriptions"],"summary":"Activate","operationId":"activate_subscriptions_activate_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subscription"}}}}},"security":[{"HTTPBearer":[]}]}},"/subscriptions/cancel":{"post":{"tags":["Subscriptions"],"summary":"Cancel","operationId":"cancel_subscriptions_cancel_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Subscription"}}}}},"security":[{"HTTPBearer":[]}]}},"/collection-slots/me":{"get":{"tags":["CollectionSlots"],"summary":"Get Me","operationId":"get_me_collection_slots_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionSlot"}}}}},"security":[{"HTTPBearer":[]}]},"put":{"tags":["CollectionSlots"],"summary":"Put Me","operationId":"put_me_collection_slots_me_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionSlot"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CollectionSlot"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/collections":{"post":{"tags":["Collections"],"summary":"Create Collection","operationId":"create_collection_collections_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCollectionRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/collections/me":{"get":{"tags":["Collections"],"summary":"List My Collections","operationId":"list_my_collections_collections_me_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Pagesize"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/collections/{id}/cancel":{"patch":{"tags":["Collections"],"summary":"Cancel Collection","operationId":"cancel_collection_collections__id__cancel_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/claims/submit":{"post":{"tags":["Claims"],"summary":"Submit Claim","operationId":"submit_claim_claims_submit_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_submit_claim_claims_submit_post"}}},"required":true},"responses":{"202":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/return-points":{"get":{"tags":["ReturnPoints"],"summary":"List Return Points Endpoint","operationId":"list_return_points_endpoint_return_points_get","parameters":[{"name":"chain","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Chain"}},{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Q"}},{"name":"near","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"lat,lng","title":"Near"},"description":"lat,lng"},{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"default":1,"title":"Page"}},{"name":"pageSize","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":100,"title":"Pagesize"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReturnPointsResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/simulate/return":{"post":{"tags":["Simulator"],"summary":"Simulate Return","operationId":"simulate_return_simulate_return_post","parameters":[{"name":"Idempotency-Key","in":"header","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Idempotency-Key"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["Health"],"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/healthz":{"get":{"tags":["Health"],"summary":"Healthz","operationId":"healthz_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/ping":{"get":{"tags":["Admin"],"summary":"Ping","operationId":"ping_admin_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/admin/collections/{id}/status":{"patch":{"tags":["Admin"],"summary":"Update Collection Status","operationId":"update_collection_status_admin_collections__id__status_patch","security":[{"HTTPBearer":[]}],"parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateStatusRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AuthLoginRequest":{"properties":{"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["email","password"],"title":"AuthLoginRequest"},"AuthTokenResponse":{"properties":{"accessToken":{"type":"string","title":"Accesstoken"},"tokenType":{"type":"string","title":"Tokentype"},"expiresIn":{"type":"integer","title":"Expiresin"},"user":{"additionalProperties":true,"type":"object","title":"User"}},"type":"object","required":["accessToken","tokenType","expiresIn","user"],"title":"AuthTokenResponse"},"Body_submit_claim_claims_submit_post":{"properties":{"image":{"type":"string","format":"binary","title":"Image"},"note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Note"}},"type":"object","required":["image"],"title":"Body_submit_claim_claims_submit_post"},"CollectionSlot":{"properties":{"weekday":{"type":"integer","title":"Weekday"},"startTime":{"type":"string","title":"Starttime"},"endTime":{"type":"string","title":"Endtime"},"preferredReturnPointId":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Preferredreturnpointid"}},"type":"object","required":["weekday","startTime","endTime"],"title":"CollectionSlot"},"CreateCollectionRequest":{"properties":{"scheduledAt":{"type":"string","title":"Scheduledat"},"returnPointId":{"type":"integer","title":"Returnpointid"},"bagCount":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Bagcount","default":1},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["scheduledAt","returnPointId"],"title":"CreateCollectionRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ReturnPoint":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"type":{"type":"string","title":"Type"},"eircode":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Eircode"},"retailer":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Retailer"},"lat":{"type":"number","title":"Lat"},"lng":{"type":"number","title":"Lng"}},"type":"object","required":["id","name","type","lat","lng"],"title":"ReturnPoint"},"ReturnPointsResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/ReturnPoint"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"}},"type":"object","required":["items","total"],"title":"ReturnPointsResponse"},"Subscription":{"properties":{"status":{"type":"string","title":"Status"},"planCode":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Plancode"},"startDate":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Startdate"},"endDate":{"anyOf":[{"type":"string","format":"date"},{"type":"null"}],"title":"Enddate"}},"type":"object","required":["status"],"title":"Subscription"},"Transaction":{"properties":{"id":{"type":"integer","title":"Id"},"ts":{"type":"string","format":"date-time","title":"Ts"},"kind":{"type":"string","title":"Kind"},"amountCents":{"type":"integer","title":"Amountcents"},"note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Note"}},"type":"object","required":["id","ts","kind","amountCents"],"title":"Transaction"},"UpdateStatusRequest":{"properties":{"status":{"type":"string","title":"Status"}},"type":"object","required":["status"],"title":"UpdateStatusRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"WalletBalanceResponse":{"properties":{"balanceCents":{"type":"integer","title":"Balancecents"},"lastUpdated":{"type":"string","format":"date-time","title":"Lastupdated"}},"type":"object","required":["balanceCents","lastUpdated"],"title":"WalletBalanceResponse"},"WalletHistoryResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/Transaction"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"page":{"type":"integer","title":"Page"},"pageSize":{"type":"integer","title":"Pagesize"}},"type":"object","required":["items","total","page","pageSize"],"title":"WalletHistoryResponse"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}